==> scRNAseq_Analysis/03_scRNAseq.Seurat.LymphoidAnalysis.Rmd <==
```{r}
library(Seurat)
library(Matrix)
library(dplyr)
library(cowplot)
require(data.table)
library(harmony)
library("ggplot2")
library(ggpubr)
library(RColorBrewer)
library(reshape2)


options(bitmapType='cairo') 
prefix = "Seurat.MetsGBM.CombinedAnalysis.Apr2022.Lymphoid"
infile = "Seurat.MetsGBM.CombinedAnalysis.Apr2022.MinD0.1.sprd5.NN20.seed122.ep1000.rds"
id_file = "Seurat.MetsGBM.CombinedAnalysis.Apr2022.MinD0.1.sprd5.NN20.seed122.ep1000.Lymphoid.ID.txt"

```

```{r}
ids <- as.character(read.table(id_file, header = FALSE)[,1])
data.all=readRDS(file=infile)
data.combined = subset(data.all, cells = ids)
rm(data.all)

data.combined@meta.data = data.combined@meta.data[!colnames(data.combined@meta.data) %in% c("integrated_snn_res.0.3","integrated_snn_res.0.5","integrated_snn_res.0.7","seurat_clusters")]

ob.list<-SplitObject(data.combined,split.by= "condition")

ob.list.gbm.rec <- SplitObject(ob.list$GBM.rec,split.by= "ID")
gbm.merge <- merge(
  x = ob.list.gbm.rec[[1]],
  y = unlist(ob.list.gbm.rec[2:length(ob.list.gbm.rec)])
)
gbm.merge@assays$RNA@var.features <- SelectIntegrationFeatures(ob.list.gbm.rec)


ob.list.gbm.pembro <- SplitObject(ob.list$GBM.pembro,split.by= "ID")
gbm.pembro.merge <- merge(
  x = ob.list.gbm.pembro[[1]],
  y = unlist(ob.list.gbm.pembro[2:length(ob.list.gbm.pembro)])
)
gbm.pembro.merge@assays$RNA@var.features <- SelectIntegrationFeatures(ob.list.gbm.pembro)


ob.list.met <- SplitObject(ob.list$met,split.by= "ID")
met.merge <- merge(
  x = ob.list.met[[1]],
  y = unlist(ob.list.met[2:length(ob.list.met)])
)
met.merge@assays$RNA@var.features <- SelectIntegrationFeatures(ob.list.met)


ob.list.met.pembro <- SplitObject(ob.list$met.pembro,split.by= "ID")
met.pembro.merge <- merge(
  x = ob.list.met.pembro[[1]],
  y = unlist(ob.list.met.pembro[2:length(ob.list.met.pembro)])
)
met.pembro.merge@assays$RNA@var.features <- SelectIntegrationFeatures(ob.list.met.pembro)

ob.list = list(gbm.merge, gbm.pembro.merge, met.merge, met.pembro.merge)

rm(ob.list.gbm.rec)
rm(ob.list.gbm.pembro)
rm(ob.list.met)
rm(ob.list.met.pembro)
rm(data.combined)

gc()

```

#Run integration, scaling and PCA
```{r}
k.filterMax <- min(200, min(sapply(ob.list, ncol)))
anchors <- FindIntegrationAnchors(object.list = ob.list, k.filter = k.filterMax, anchor.features = 2000)
data.combined <- IntegrateData(anchorset = anchors, k.weight = k.filterMax)

DefaultAssay(object = data.combined) <- "integrated"

data.combined <- ScaleData(object = data.combined, verbose = TRUE, vars.to.regress = c("percent.mt","percent.ribo","nFeature_RNA","nCount_RNA","S.Score","G2M.Score"))

gc()

data.combined <- RunPCA(object = data.combined, npcs = 30, verbose = TRUE)

rm(anchors)
gc()

```

#Run harmony
```{r}

data.combined <- RunHarmony(data.combined, group.by.vars="Dataset",assay.use = "integrated", plot_convergence = TRUE)
harmony_embeddings <- Embeddings(data.combined, 'harmony')

gc()
```


#Run UMAP for cluster visualization
```{r}
minDist = 0.1
spread1 = 5
NN = 20
seedUse = 122
epoch = 1000

prefix1 = paste(prefix,".MinD",minDist,".sprd",spread1,".NN",NN,".seed",seedUse,".ep",epoch,sep="")

data.combined <- RunUMAP(object = data.combined, reduction = "harmony", dims = 1:30,min.dist = minDist, spread=spread1, n.neighbors=NN, seed.use = seedUse, n.epochs=epoch)

outfile = paste(prefix1,"_UMAP", ".nUMI.png",sep="")
png(filename = outfile,width = 800, height = 800)
FeaturePlot(object = data.combined, features = c("nCount_RNA"), min.cutoff = "q9", pt.size = 2) #umi plot
dev.off()

outfile = paste(prefix1,"_UMAP",".nGene.png",sep="")
png(filename = outfile,width = 800, height = 800)
FeaturePlot(object = data.combined, features = c("nFeature_RNA"), min.cutoff = "q9", pt.size = 2) #umi plot
dev.off()

outfile = paste(prefix1,"_UMAP",".pctRibo.png",sep="")
png(filename = outfile,width = 800, height = 800)
FeaturePlot(object = data.combined, features = c("percent.ribo"), min.cutoff = "q9", pt.size = 2) #umi plot
dev.off()

outfile = paste(prefix1,"_UMAP",".pctMito.png",sep="")
png(filename = outfile,width = 800, height = 800)
FeaturePlot(object = data.combined, features = c("percent.mt"), min.cutoff = "q9", pt.size = 2) #umi plot
dev.off()

data.combined@meta.data$Dataset = factor(data.combined@meta.data$Dataset, levels = 
                                           c("GBM.2021","Mets.2022","Werb.2022","Feldman"))


outfile = paste(prefix1,"_UMAP.dataset.png",sep="")
png(filename = outfile,width = 800, height = 800)
DimPlot(object = data.combined, reduction = "umap", group.by = "Dataset", label = TRUE, label.size = 8, pt.size = 0.5)
dev.off()

outfile = paste(prefix1,"_UMAP.dataset.split.png",sep="")
png(filename = outfile,width = 1600, height = 800)
DimPlot(object = data.combined, reduction = "umap", group.by = "Dataset", split.by = "Dataset", label = FALSE, label.size = 8, pt.size = 0.5)
dev.off()


data.combined@meta.data$condition = factor(data.combined@meta.data$condition, levels = 
                                             c("GBM.rec","GBM.pembro","met","met.pembro"))
color_condition <- c("cyan","blue","pink","red")

outfile = paste(prefix1,"_UMAP.condition.png",sep="")
png(filename = outfile,width = 800, height = 800)
DimPlot(object = data.combined, reduction = "umap", group.by = "condition", label = FALSE, label.size = 8, pt.size = 0.5, cols = color_condition)
dev.off()

outfile = paste(prefix1,"_UMAP.condition.split.png",sep="")
png(filename = outfile,width = 1600, height = 800)
DimPlot(object = data.combined, reduction = "umap", group.by = "condition", split.by = "condition", label = FALSE, label.size = 8, pt.size = 0.5, cols = color_condition)
dev.off()

outfile = paste(prefix1,"_UMAP.WerbCelltype.png",sep="")
png(filename = outfile,width = 800, height = 800)
DimPlot(object = data.combined, reduction = "umap", group.by = "Cell_Type", label = TRUE, label.size = 8, pt.size = 1, na.value = "lightgrey", raster=FALSE)
dev.off()

outfile = paste(prefix1,"_UMAP.Sample.png",sep="")
png(filename = outfile,width = 2000, height = 1200)
DimPlot(object = data.combined, reduction = "umap", group.by = "ID", label = TRUE, label.size = 4, pt.size = 1, na.value = "lightgrey", raster=FALSE)
dev.off()

meta.data = data.combined@meta.data
meta.data$MainCancerType = "OtherMets"
meta.data[meta.data$CancerType %in% c("Melanoma"),]$MainCancerType = "Melanoma"
meta.data[meta.data$CancerType %in% c("Recurrent"),]$MainCancerType = "GBM"

data.combined@meta.data = meta.data

meta.data = data.combined@meta.data
meta.data$MainCancerType2 = "OtherMets"
meta.data[meta.data$CancerType %in% c("Recurrent"),]$MainCancerType2 = "GBM"
meta.data[meta.data$CancerType %in% c("Melanoma"),]$MainCancerType2 = "Melanoma"
meta.data[meta.data$CancerType %in% c("Breast"),]$MainCancerType2 = "Breast"
#meta.data[meta.data$CancerType %in% c("Endometrial"),]$MainCancerType2 = "Endometrial"


data.combined@meta.data = meta.data

saveRDS(data.combined, file=paste(prefix1, "rds", sep="."))

```


#Clustering
```{r}
DefaultAssay(object = data.combined) <- "integrated"

data.combined <- FindNeighbors(object = data.combined, reduction = "harmony", dims = 1:30, nn.method = "annoy", annoy.metric = "cosine", force.recalc = TRUE)
data.combined <- FindClusters(data.combined, resolution = 0.7)

#colors = c("red","orange","yellow","green","lightgreen","blue","magenta","cyan","pink","brown","black","darkblue","darkgrey","lightgrey","maroon","lightblue")

outfile = paste(prefix1,"_UMAP", ".Res0.7.png",sep="")
png(filename = outfile,width = 800, height = 800)
DimPlot(object = data.combined, reduction = "umap", group.by = "integrated_snn_res.0.7", label = TRUE, label.size = 8, pt.size = 1)#, cols = colors[1:8])
dev.off()

saveRDS(data.combined, file=paste(prefix1, "rds", sep="."))
#data.combined = readRDS(file=paste(prefix1, ".rds", sep=""))

```

#draw umap of genes of interest ###
```{r}
DefaultAssay(data.combined) <- "RNA"

gene = c("CD3D","CD8A",
         "CD4","CD14",
         "CD52","TRAC",
         "IL7R","TCF7",
         "CCR7","CD69",
         "CXCR4","CXCR6",
         "CD28","CD226",
         "PRDM1","ENTPD1",
         "TOX","HAVCR2",
         "BATF","SLAMF6",
         "PDCD1","GNLY",
         "GZMB","PRF1",
         "GZMK","GZMH",
         "SELL","IFNG",
         "ICOS","CTLA4",
         "TIGIT","FOXP3",
         "CXCL13","NR4A2",
         "MKI67","CD79A",
         "KLRB1","FCGR3A")


outfile = paste(prefix,"_UMAP.MainCellTypes.png",sep="")
png(filename = outfile,width = 1200, height = 2000)
FeaturePlot(object = data.combined, features = gene, min.cutoff = "q9",  pt.size = 1, ncol=4, cols = c("lightgrey","red"))
dev.off()

```

#Find cluster markers
```{r}
DefaultAssay(data.combined) <- "RNA"
Idents(object = data.combined) <- "integrated_snn_res.0.7"

cluster.markers<- FindAllMarkers(object = data.combined, only.pos = TRUE, min.pct = 0.25,logfc.threshold = 0.322,test.use="MAST")
outfile = paste(prefix1,"_ClusterDiffGene.Res0.7", ".csv",sep="")
write.csv(cluster.markers, file=outfile)

saveRDS(data.combined,file=paste(prefix1, ".rds", sep=""))
#data.combined = readRDS(file=paste(prefix1, ".rds", sep=""))

#looking for DEGs between two clusters
cluster.markers<- FindMarkers(object = data.combined, ident.1 = 9, ident.2 = 0, only.pos = FALSE, min.pct = 0.25,logfc.threshold = 0.322,test.use="MAST")

outfile = paste(prefix1,"_DEGBetweenCluster0and9", ".csv",sep="")
write.csv(cluster.markers, file=outfile)

```

#assign cell types ###
```{r}
#assign each cluster cell type
Idents(object = data.combined) <- "integrated_snn_res.0.7"
current.cluster.ids <- c(0:25)
new.cluster.ids <- c("CD8-VCAM1-IFNG-Tex","Treg","CD8-ITGA1-ITGAL-Trm", #2
                     "Tumor+Tcell","CD8-CXCL13-LAG3-Tex","CD4-IL7R-CCR7-CD40LG-Tcm",#5
                    "Mito High","CTL/NK","Mphage+Tcell",#8
                    "CD8-VCAM1-IFNG-Tex","CD8-GZMK-CCL4-CCL5-Teff","CD8-GZMH-GZMA-CD52-Teff",#11
                    "CD8-IL7R-CD69-Tm","CTL/NK-XCL1-XCL2","Lymphoid.Cycling", #14
                    "CD8-TCF7-CD226-Tprog.exh","CD8-ISG High","Tumor+Tcell", #17
                   "IL7R-KLRB1-Tm","CD4-CXCL13-CTLA4-Tex","Tumor+Tcell", #20
                   "Tumor+Tcell","Mito High","Mast cell",#23
                   "Tumor+Tcell","Tumor+Tcell") #25
        

Idents(object = data.combined) <- plyr::mapvalues(x = Idents(object = data.combined), from = current.cluster.ids, to = new.cluster.ids)

data.combined@meta.data$celltype <- Idents(object = data.combined)
data.combined@meta.data$celltype <- factor(data.combined@meta.data$celltype,
                                           levels = c("CD8-IL7R-CD69-Tm",
                                                      "CD8-ITGA1-ITGAL-Trm",
                                                      "CD8-TCF7-CD226-Tprog.exh",
                                                      "CD8-ISG High",
                                                      "CD8-GZMK-CCL4-CCL5-Teff",
                                                      "CD8-GZMH-GZMA-CD52-Teff",
                                                      "CD8-VCAM1-IFNG-Tex",
                                                      "CD8-CXCL13-LAG3-Tex",
                                                      "CD4-IL7R-CCR7-CD40LG-Tcm",
                                                      "CD4-CXCL13-CTLA4-Tex",
                                                      "Treg",
                                                      "IL7R-KLRB1-Tm",
                                                      "CTL/NK",
                                                      "CTL/NK-XCL1-XCL2",
                                                      "Lymphoid.Cycling",
                                                      "Mito High",
                                                      "Tumor+Tcell",
                                                      "Mphage+Tcell",
                                                      "Mast cell"))

cols = c(brewer.pal(12, name="Paired"),"#FFD92F","#E5C494","#E78AC3","#DCDCDC","#666666","#778899","black")
cols[6] <- "maroon"
#brewer.pal(8, name="Dark2")


outfile = paste(prefix1,"_UMAP.Celltype.png",sep="")
png(filename = outfile,width = 1000, height = 800)
DimPlot(object = data.combined, reduction = "umap", group.by = "celltype", label = TRUE, label.size = 5, pt.size = 1, repel=TRUE, cols = cols, raster=FALSE)
dev.off()

outfile = paste(prefix1,"_UMAP.Celltype.noLabel.tiff",sep="")
tiff(file=outfile, width=4000, height=2800, res=600)
DimPlot(object = data.combined, reduction = "umap", group.by = "celltype", label = FALSE, label.size = 8, pt.size = 0.2, cols = cols)
dev.off()

saveRDS(data.combined,file=paste(prefix1, ".rds", sep=""))
#data.combined = readRDS(file=paste(prefix1, ".rds", sep=""))
```

#Find celltype markers
```{r}
DefaultAssay(data.combined) <- "RNA"
Idents(object = data.combined) <- "celltype"

cluster.markers<- FindAllMarkers(object = data.combined, only.pos = TRUE, min.pct = 0.25,logfc.threshold = 0.322,test.use="MAST")

cluster.markers$DeltaPerLogFC = cluster.markers$avg_log2FC*(cluster.markers$pct.1-cluster.markers$pct.2)
  
outfile = paste(prefix1,"_CelltypeDiffGene", ".csv",sep="")
write.csv(cluster.markers, file=outfile)

```


#draw boxplots of genes of interest and output celltype-based average expressions ###
```{r}
library(ggplot2)
library(reshape)
library(dplyr)
library(ggpubr)
library(scales)

gene = c("CD3D","CD8A","CD8B","CD4",
         "IL7R","CCR7","SELL","TCF7",
         "CD40LG","CD28","CD226","TNFRSF9",
         "CD69","ITGA1","ITGAL","CXCR6",
         "ISG15","IFI6","STAT1","CD52",
         "GZMB","GZMA","GZMK","GZMH",
         "CCL4","CCL5","VCAM1","CXCR4",
         "IFNG","CXCL13","PDCD1","CTLA4",
         "ENTPD1","NR4A2","TOX","TIGIT",
         "LAG3","HAVCR2","FOXP3","IL2RA",
         "GNLY","PRF1","KLRF1","KLRB1",
         "XCL1","XCL2","MKI67","MALAT1",
         "KRT19","SOX4","CD14","TPSB2")

Celltypes = c("CD8-IL7R-CD69-Tm",
              "CD8-GZMH-GZMA-CD52-Teff",
              "CD8-GZMK-CCL4-CCL5-Teff",
              "CD8-ISG High",
              "CD8-ITGA1-ITGAL-Trm",
              "CD8-TCF7-CD226-Tprog.exh",
              "CD8-VCAM1-IFNG-Tex",
              "CD8-CXCL13-LAG3-Tex",
              "CD4-IL7R-CCR7-CD40LG-Tcm",
              "CD4-CXCL13-CTLA4-Tex",
              "Treg",
              "IL7R-KLRB1-Tm",
              "CTL/NK",
              "CTL/NK-XCL1-XCL2",
              "Lymphoid.Cycling")#,
              #"Mito High",
              #"Tumor+Tcell",
              #"Mphage+Tcell",
              #"Mast cell")

SubsetData = subset(data.combined, subset = (celltype %in% Celltypes))
SubsetData$celltype = factor(SubsetData$celltype, levels = Celltypes)

data = SubsetData@assays$RNA@data
expr_gene = data[gene,]
expr_gene = as.data.frame(t(expr_gene))

expr_gene$condition = SubsetData@meta.data$celltype
expr_gene$condition = factor(SubsetData@meta.data$celltype,levels=Celltypes)

colnames(expr_gene) = c(gene,"condition")

colors = cols
#colors = hue_pal()(19)

ncol = 4
nrow1 = length(gene)/ncol
height1 = nrow1*180
width1 = length(levels(expr_gene$condition))*20*ncol

outfile = paste(prefix,"GeneBoxplot.ByCelltype.png",sep="_")
png(filename = outfile,width = width1, height = height1)
panels=list()
l=length(colnames(expr_gene))
i=1
for(i in 1:length(gene)){
#for(i in 1:2){
  gene2 = colnames(expr_gene)[i]
  data2 = expr_gene[,c(i,l)]
  colnames(data2)=c("normExpr","condition")
  
  #geom_boxplot(outlier.colour="black", outlier.shape=16,
             #outlier.size=2, notch=FALSE)
 
  y = ggplot(data=data2, aes(x=condition,y=normExpr,fill=condition)) +
       #geom_violin(trim = FALSE)+
       geom_boxplot(outlier.colour="black",notch=TRUE)+
        scale_fill_manual(values=colors)+
      theme(
        axis.text.x = element_blank(),
        axis.text.y = element_text(size=28),
        axis.ticks.x = element_blank(),
        axis.title = element_blank(),
        legend.position = "none",
        plot.title=element_text(face = "italic",size=32))+
      ggtitle(gene2)
    
      panels[[i]]=y
  
}
ggarrange(plotlist=panels,ncol=ncol,nrow=nrow1)
dev.off()


##compuate celltype/cluster-based average expressions
Idents(object = SubsetData) <- "celltype"
AverageExpr <- AverageExpression(SubsetData,
                                 assays = "RNA",
                                 features = gene,
                                 group.by= "ident",
                                 return.seurat = TRUE)

outfile = paste(prefix,"_AverageExprByCelltype.txt",sep="")
write.table(AverageExpr@assays$RNA@scale.data,file=outfile,sep="\t")

genelist = c("CD3D","CD8A","CD8B","CD4",
          "IL7R","CCR7","CD69","CD40LG",
          "GZMA","GZMK","GZMH","GZMB",
          "ITGA1","ITGAL",
          "TCF7","CD226",
          "IFNG","PDCD1",
          "CTLA4","TOX","TIGIT","LAG3","HAVCR2","CXCL13",
          "FOXP3",
          "KLRB1","XCL1","XCL2",
          "MKI67")

#genelist = c("CD52")
# outfile = paste(prefix,"Dotplot.SelectedGene.png",sep=".")
# png(outfile, height = 2000, width = 4000)
outfile = paste(prefix,"Dotplot.SelectedGene.tiff",sep=".")
tiff(file=outfile, width=4000, height=2000, res=400)
DotPlot(SubsetData, features = genelist)
dev.off()

```


## Write the files into separate populations for subset analysis
```{r}

Exprlist1 = data.combined@assays$RNA@counts["CD14",]+data.combined@assays$RNA@counts["CD68",]
Exprlist2 = data.combined@assays$RNA@counts["CD8A",]+data.combined@assays$RNA@counts["CD8B",]
Exprlist3 = data.combined@assays$RNA@counts["CD3D",]+data.combined@assays$RNA@counts["CD3E",]+data.combined@assays$RNA@counts["CD3G",]
Exprlist4 = data.combined@assays$RNA@counts["CD4",]

## no myeloid only and no CD4 (CD8 only)
ids = names(Exprlist1[Exprlist1 < 1 & Exprlist2 >= 1 & Exprlist3 >= 1 & Exprlist4 < 1])
fileConn<-file(paste(prefix1,".CD8.ID.txt",sep=""))
writeLines(ids, fileConn)

data.sub = subset(data.combined, cells = ids)
DefaultAssay(data.sub) <- "RNA"
outfile = paste(prefix,"_UMAP",".CD8.png",sep="")
png(filename = outfile,width = 1600, height = 800)
FeaturePlot(object = data.sub, features = c("CD8A","CD8B"), min.cutoff = 0, pt.size = 1, cols=c('lightgrey', 'red'))
dev.off()


ids = names(Exprlist1[Exprlist1 < 1 & Exprlist2 < 1 & Exprlist3 >= 1 & Exprlist4 >= 1])

fileConn<-file(paste(prefix1,".CD4.ID.txt",sep=""))
writeLines(ids, fileConn)

data.sub = subset(data.combined, cells = ids)
DefaultAssay(data.sub) <- "RNA"
outfile = paste(prefix,"_UMAP",".CD4.png",sep="")
png(filename = outfile,width = 1600, height = 800)
FeaturePlot(object = data.sub, features = c("CD3D","CD4"), min.cutoff = 0, pt.size = 1, cols=c('lightgrey', 'red'))
dev.off()

```

#calculate each cell type's number in each sample ###
```{r}
## Enumerate per sample
totalNum=c(table(data.combined@meta.data$ID)[[1]])
Group = names(table(data.combined@meta.data$ID))
table_sub = data.frame()
table_sub = subset(data.combined@meta.data, ID==Group[1])
percentages=as.data.frame((table(table_sub$celltype)))$Freq
table1 = percentages

#m1 = data.combined@meta.data

for(i in 2:length(Group)){
  id = Group[i]
  n_total = table(data.combined@meta.data$ID)[[i]]
  totalNum=c(totalNum,n_total)
  
  table_sub = data.frame()
  table_sub = subset(data.combined@meta.data, ID==id)
  
  percentages=as.data.frame((table(table_sub$celltype)))$Freq
  table1=as.data.frame(rbind(table1,percentages))
}

table2 = as.data.frame(cbind(table1,Total=totalNum))
rownames(table2)=Group
colnames(table2)=c(levels(data.combined@meta.data$celltype),"Total")
table2$Lymphoid = table2$Total-table2$'Tumor+Tcell'-table2$'Mphage+Tcell'-table2$'Mast cell'-table2$'Mito High'

sampleMap = unique(data.combined@meta.data[,c("ID","condition","CancerType","MainCancerType")])
rownames(sampleMap) = sampleMap$ID
sampleMap = sampleMap[Group,]
table2$condition = sampleMap$condition
table2$CancerType = sampleMap$CancerType
table2$MainCancerType = sampleMap$MainCancerType

outfile = paste(prefix,"_CelltypeByID.txt",sep="")
write.table(table2,file=outfile,sep="\t")

```


## Get DEGs for all/subset of T cells
```{r}
DefaultAssay(data.combined) <- "RNA"

Group = names(table(data.combined@meta.data$ID))
sampleMap = unique(data.combined@meta.data[,c("ID","condition","CancerType")])
rownames(sampleMap) = sampleMap$ID
sampleMap = sampleMap[Group,]

Celltypes = c("CD8-IL7R-CD69-Tm",
              "CD8-ITGA1-ITGAL-Trm",
              "CD8-TCF7-CD226-Tprog.exh",
              "CD8-ISG High",
              "CD8-GZMK-CCL4-CCL5-Teff",
              "CD8-GZMH-GZMA-CD52-Teff",
              "CD8-VCAM1-IFNG-Tex",
              "CD8-CXCL13-LAG3-Tex",
              "CD4-IL7R-CCR7-CD40LG-Tcm",
              "CD4-CXCL13-CTLA4-Tex",
              "Treg",
              "IL7R-KLRB1-Tm",
              "CTL/NK",
              "CTL/NK-XCL1-XCL2",
              "Lymphoid.Cycling")#,
              #"Mito High")#,
              #"Tumor+Tcell",
              #"Mphage+Tcell",
              #"Mast cell")

SubsetData = subset(data.combined, subset = (celltype %in% Celltypes))

#sample the same number of cells from every sample to avoid patient-specific effects
IDlist = c()
nCell = 50
set.seed(123)
for(i in 1:length(Group)){
  ids_sub = c()
  
  subdata = subset(SubsetData,ID==Group[i])
  ids = colnames(subdata)
  
  if(length(ids) < nCell){
     ids_sub = ids
  }else{
     ids_sub = sample(ids, nCell)
  }
  IDlist = c(IDlist,ids_sub)
}


## get DEG in met vs GBM samples ####
conditions = c("GBM.rec","met")
TreatmentSubset = subset(data.combined[,IDlist],condition %in% conditions)
Idents(object = TreatmentSubset) <- "condition"

markers <- FindAllMarkers(TreatmentSubset,min.pct = 0.25, logfc.threshold = 0.322,only.pos = TRUE)
markers <- as.data.frame(markers)
markers$diff = markers$pct.1 - markers$pct.2
markers$DeltaPerMultiLog2FC = markers$diff*markers$avg_log2FC

outfile = paste(prefix,"DEG","BrMvsGBM.NoPembro","NoDoubletNoMast.tsv",sep=".")
write.table(markers, file=outfile, sep="\t" )


## get DEG in BrM.ICB vs BrM samples ####
conditions = c("met","met.pembro")
TreatmentSubset = subset(data.combined[,IDlist],condition %in% conditions)

Idents(object = TreatmentSubset) <- "condition"
markers <- FindAllMarkers(TreatmentSubset,min.pct = 0.25, logfc.threshold = 0.322,only.pos = TRUE)
markers <- as.data.frame(markers)
markers$diff = markers$pct.1 - markers$pct.2
markers$DeltaPerMultiLog2FC = markers$diff*markers$avg_log2FC

outfile = paste(prefix,"DEG","MetsICBvsMets","NoDoubletNoMast.tsv",sep=".")
write.table(markers, file=outfile, sep="\t" )


## get DEG in met vs GBM samples ####
conditions = c("GBM.rec","GBM.pembro")
TreatmentSubset = subset(data.combined[,IDlist],condition %in% conditions)
Idents(object = TreatmentSubset) <- "condition"

markers <- FindAllMarkers(TreatmentSubset,min.pct = 0.25, logfc.threshold = 0.322,only.pos = TRUE)
markers <- as.data.frame(markers)
markers$diff = markers$pct.1 - markers$pct.2
markers$DeltaPerMultiLog2FC = markers$diff*markers$avg_log2FC

outfile = paste(prefix,"DEG","GBMICBvsGBM","NoDoubletNoMast.tsv",sep=".")
write.table(markers, file=outfile, sep="\t" )

## get DEG in melanoma mets vs other mets - untreated ####
conditions = c("met")
TreatmentSubset = subset(data.combined[,IDlist],condition %in% conditions)
Idents(object = TreatmentSubset) <- "CancerType"

markers <- FindAllMarkers(TreatmentSubset,min.pct = 0.25, logfc.threshold = 0.322,only.pos = TRUE)
markers <- as.data.frame(markers)
markers$diff = markers$pct.1 - markers$pct.2
markers$DeltaPerMultiLog2FC = markers$diff*markers$avg_log2FC

outfile = paste(prefix,"DEG","MetsCancerType","NoDoubletNoMast.tsv",sep=".")
write.table(markers, file=outfile, sep="\t" )


```

## get sample-level averaged gene expression boxplot/violin plot
```{r}
SubsetData = subset(data.combined, subset = (celltype %in% Celltypes))

Group = names(table(SubsetData@meta.data$ID))
sampleMap = unique(SubsetData@meta.data[,c("ID","condition","CancerType","MainCancerType","MainCancerType2")])
rownames(sampleMap) = sampleMap$ID
sampleMap = sampleMap[Group,]

#gene sample-wise gene expression level
Idents(SubsetData) <- "ID"
expr_avg <- AverageExpression(SubsetData,
                              assays = "RNA",
                              group.by = "ident",
                              #add.ident = "condition",
                              slot = "data")
expr_avg = as.data.frame(expr_avg$RNA)
expr_avg = expr_avg[,Group]

Genes = c("TRAC","TRBC1","CD52","GZMB","B2M","CD27",#Met vs GBM.rec, no pembro
          "CTLA4","TIGIT","ICOS","TNFRSF18","CXCR4","BATF",#Met vs GBM.rec, no pembro
          "PDCD1","CXCL13","TOX","LAG3","HAVCR2","TIGIT","TNFRSF14","PRDM1","NR4A2", 
          "IFITM1","HLA-DRB5","CD81","CCL5","CD8A","CD8B",
          "STAT1","JAK1","JAK3","STAT3","IRF1",
          "GZMH","GZMK","CD27","IFNG","TNFRSF9","MKI67","CD38","ITGA4","CCR7")

expr_gene = expr_avg[unique(Genes),sampleMap$ID]
expr_gene = as.data.frame(t(expr_gene))

expr_gene$CancerType = sampleMap$CancerType
expr_gene$condition = sampleMap$condition
expr_gene$MainCancerType = sampleMap$MainCancerType
expr_gene$MainCancerType2 = sampleMap$MainCancerType2

colnames(expr_gene) = c(unique(Genes),"type","condition","MainType","MainType2")

outfile = paste(prefix,"_GeneExprByID.txt",sep="")
write.table(expr_gene,file=outfile,sep="\t")


```

## plot per-celltype-based gene expression/gene set score boxplot/violin plot
```{r}

Celltypes = c("CD8-IL7R-CD69-Tm",
              "CD8-GZMH-GZMA-CD52-Teff",
              "CD8-GZMK-CCL4-CCL5-Teff",
              "CD8-ISG High",
              "CD8-ITGA1-ITGAL-Trm",
              "CD8-TCF7-CD226-Tprog.exh",
              "CD8-VCAM1-IFNG-Tex",
              "CD8-CXCL13-LAG3-Tex",
              "CD4-IL7R-CCR7-CD40LG-Tcm",
              "CD4-CXCL13-CTLA4-Tex",
              "Treg",
              "IL7R-KLRB1-Tm",
              "CTL/NK",
              "CTL/NK-XCL1-XCL2",
              "Lymphoid.Cycling")#,
              #"Mito High",
              #"Tumor+Tcell",
              #"Mphage+Tcell",
              #"Mast cell")

condition_sub = c("GBM.rec","met")
#condition_sub = c("met","met.pembro")
#condition_sub = c("GBM.rec","GBM.pembro","met","met.pembro")

SubsetData = subset(data.combined, celltype %in% Celltypes & condition %in% condition_sub)

Group = names(table(SubsetData@meta.data$ID))

#sample the same number of cells from every sample to avoid patient-specific effects
IDlist = c()
nCell = 50
set.seed(123)
for(i in 1:length(Group)){
  ids_sub = c()
  
  subdata = subset(SubsetData,ID==Group[i])
  ids = colnames(subdata)
  
  if(length(ids) < nCell){
     ids_sub = ids
  }else{
     ids_sub = sample(ids, nCell)
  }
  IDlist = c(IDlist,ids_sub)
}

TreatmentSubset = data.combined[,IDlist]
DefaultAssay(TreatmentSubset)<-"RNA"

#Genes = c("TRAC","TRBC1","CD52","GZMB","PDCD1",
#          "CTLA4","TIGIT","ICOS","MKI67","CXCR4")#Met vs GBM.rec, no pembro

#Genes = c("TNFRSF9","HAVCR2","MKI67","CXCL13","IFNG","CD38","TIGIT","PDCD1",
#          "LAG3","CTLA4")#melanoma vs other mets, no pembro

#Genes = c("CD8A","CD8B","GZMK","GZMH","CCL5","IRF1",
#          "CXCL13","PRDM1","PDCD1","TOX","HAVCR2","NR4A2")#Met vs mets.pembro

#Genes = c("CXCR4","CXCR6")
Genes = c("TRAC","TRBC1","GZMB","CTLA4","TIGIT","ICOS")

#for genes ###
expr_gene = TreatmentSubset@assays$RNA@data[Genes,]
expr_gene = as.data.frame(t(expr_gene))

#for gene sets ###
#expr_gene = TreatmentSubset@meta.data[,Genes]

expr_gene$type = TreatmentSubset@meta.data$celltype
expr_gene$condition = TreatmentSubset@meta.data$condition
expr_gene$MainType = TreatmentSubset@meta.data$MainCancerType

expr_gene <- droplevels(expr_gene)
expr_gene$type = factor(expr_gene$type, levels = Celltypes)
expr_gene$condition = factor(expr_gene$condition, levels = condition_sub)


#colnames(expr_gene) = c(Genes,"type","condition","MainType")

colors = c("cyan","pink")
#colors = c("pink","red")
#colors = c("cyan","#F0027F","#666666")
#colors = c("cyan","orange","red")
#colors = c("cyan","pink")
#colors = c("cyan","blue","pink","red")

ncol = 1  
nrow1 = length(Genes)/ncol
width1 = length(Celltypes)*80*ncol
height1 = nrow1*180

outfile = paste(prefix,"DEG.Boxplot.ByCelltype.CTLA4.png",sep=".") #gene expr boxplot in each cell type
png(filename = outfile,width = width1, height = height1)
panels=list()
l=length(colnames(expr_gene))

for(j in 1:length(Genes)){
#for(i in 1:2){
  gene2 = colnames(expr_gene)[j]
  data2 = expr_gene[,c(j,l-2,l-1,l)]
  colnames(data2)=c("normExpr","type","condition","MainType")
 
  y = ggplot(data=data2, aes(x=type,y=normExpr,fill=condition)) +
       #geom_violin(trim = FALSE)+
      geom_boxplot(outlier.colour=alpha("grey",alpha = 0),notch=TRUE)+
      geom_point(aes(fill = condition), size = 0.5, shape = 21, position = position_jitterdodge())+
        scale_fill_manual(values=rep(colors,length(levels(expr_gene$type))))+
      theme(
        #axis.text.x = element_blank(),
        axis.text.x = element_text(size=8),
        axis.text.y = element_text(size=28),
        axis.ticks.x = element_blank(),
        axis.title = element_blank(),
        legend.position = "right",
        plot.title=element_text(face = "italic",size=36))+
      ggtitle(gene2)
    
      panels[[j]]=y
  
}
ggarrange(plotlist=panels,ncol=ncol,nrow=nrow1)
dev.off()

### pairwise wilcox test ###########################
for(j in 1:length(Genes)){
  gene2 = colnames(expr_gene)[j]
  print(gene2)
  
  for(i in 1:length(Celltypes)){
    data2 = expr_gene[,c(j,l-2,l-1,l)]
    colnames(data2)=c("normExpr","type","condition","MainType")
    data2.sub = data2[data2$type %in% c(Celltypes[i]) & data2$condition %in% condition_sub,]
    data2.sub <- droplevels(data2.sub)
    t <- wilcox.test(normExpr ~ condition, data = data2.sub)
    
    print(Celltypes[i])
    print(t$p.value)
  }
}



```


## annot the umap with gliph2 annotated TCRs
```{r}
#read in scTCR GLIPH2 output with manual annotation (Cancer, NonCancer, NoAnnot)
infile1 = "GLIPH2/scTCR.GLIPH2.Annotation.txt" #this is the annotated GLIPH resultsfor scTCR

infile2 = "GLIPH2/TRUST4.GLIPH2.Annotation.txt" #this is the annotated GLIPH results for TRUST4


#read in the combined scTCR table generated for the STARTRAC analysis
TCRab_file1 = "/home/hugolab/DataRAID/BrainMets.April2022/scTCR/Seurat.MetsGBM.CombinedAnalysis.Apr2022.Lymphoid.filtered_contig_annotations.pairedOnly.csv"

TCRab_file2 = "/home/hugolab/DataRAID/BrainMets.April2022/scTCR/MetsGBM_PBMCandTIL.Lymphoid.Feb.2021.TRB.abT.CellBarcode.TRUST4.tsv"


TCRGLIPH1 = read.table(infile1,sep="\t",header=TRUE) 
TCRGLIPH2 = read.table(infile2,sep="\t",header=TRUE)
TCRab1 <- read.csv(TCRab_file1)
TCRab2 <- read.table(TCRab_file2,sep="\t",header=TRUE)

TCRGLIPH = rbind(TCRGLIPH1,TCRGLIPH2) #combine GLIPH annot for scTCR and TRUST4

## when annotate the TCR with GLIPH annot, only use the TRB sequence
cc = colsplit(string=TCRab1$CDR3s_aa, pattern=";", names=c("Part1", "Part2"))
TCRab1$CDR3_aa.TCRb <- sub(".*:", "", cc$Part2) #get only the b chain to match with GLIPH2 output

TCRab2$newBarcode <- paste(sub("\\_.*", "", rownames(TCRab2)),TCRab2$sampleID, sep="_")

TCRab2.sub = TCRab2[,c("newBarcode","clono1")]
colnames(TCRab2.sub) = c("newBarcode","CDR3_aa.TCRb")

TCRab = rbind(TCRab2.sub, TCRab1[,c("newBarcode","CDR3_aa.TCRb")]) #combine TCR tables of scTCR and TRUST4

Merge.TCR.GLIPH <- merge(TCRab,TCRGLIPH,by.x="CDR3_aa.TCRb",by.y="CDR3s_aa",all.x=TRUE)
#Merge.meta.data.GLIPH$Annot.y[is.na(Merge.meta.data.GLIPH$Annot.y)]<-"2-NonAnnot"

GLIPHAnnot_Agg = aggregate(Merge.TCR.GLIPH,
                           by=list(Merge.TCR.GLIPH$newBarcode),
                           FUN=first) #remove duplicate rows

rownames(GLIPHAnnot_Agg) = GLIPHAnnot_Agg$Group.1

#rename the meta.data barcodes by orginal barcode with sample ID, to match the gliph2 table
meta.data = data.combined@meta.data
barcode <- sub("\\_.*", "", rownames(meta.data))
meta.data$newBarcode <- paste(barcode,meta.data$ID,sep="_")
meta.data$oldBarcode = rownames(meta.data)

#append the GLIPH annotation to meta.data
meta.data.sub = meta.data[meta.data$newBarcode %in% rownames(GLIPHAnnot_Agg),]
rownames(meta.data.sub) = meta.data.sub$newBarcode
GLIPHAnnot_Agg <- GLIPHAnnot_Agg[rownames(meta.data.sub),]

meta.data.sub = meta.data.sub[rownames(GLIPHAnnot_Agg),]
meta.data.sub$CDR3_aa.TCRb = GLIPHAnnot_Agg$CDR3_aa.TCRb
meta.data.sub$GLIPHAnnot = GLIPHAnnot_Agg$Annot
rownames(meta.data.sub) = meta.data.sub$oldBarcode

data.combined=AddMetaData(data.combined,metadata=meta.data.sub[,c("CDR3_aa.TCRb","GLIPHAnnot")])

#plot umap annotated with GLIPH annotation
data.combined@meta.data$GLIPHAnnot <- factor(data.combined@meta.data$GLIPHAnnot, 
                                          levels = c("3-Cancer","1-Viral","2-AutoImmune"))

colorsCluster = c("gold","darkgreen","black")

fp <- DimPlot(object = data.combined, reduction = "umap", group.by = "GLIPHAnnot", 
                  label = TRUE, label.size = 0, pt.size = 0.8, cols = colorsCluster, na.value =alpha("lightgrey",1))#+ na.value = alpha("lightgrey",0.1)#,cols = c(my_palette,"white"))

fp$data = rbind(fp$data[is.na(fp$data$GLIPHAnnot),],fp$data[!is.na(fp$data$GLIPHAnnot),])


outfile = paste(prefix,"_UMAP.TCRAnnot.GLIPH2Annot.png",sep="")
png(filename = outfile,width = 800, height = 800)
fp
dev.off()

## Enumerate the number of cancer and non-cancer related TCRs in each sample
Samplelist = unique(data.combined$ID)
Group = names(table(data.combined$celltype))

table_cancer <- data.frame(matrix(nrow=length(Group),ncol=length(Samplelist)))
colnames(table_cancer) = Samplelist
rownames(table_cancer) = Group

table_noncancer <- data.frame(matrix(nrow=length(Group),ncol=length(Samplelist)))
colnames(table_noncancer) = Samplelist
rownames(table_noncancer) = Group

for(i in 1:length(Samplelist)){
  sample = Samplelist[i]
  meta.data = data.combined@meta.data[data.combined$ID %in% c(sample),]
  meta.data$GLIPHAnnot = factor(meta.data$GLIPHAnnot, levels = c("3-Cancer","1-Viral","2-AutoImmune"))
  totalNum=c(table(meta.data$celltype)[[1]])
  table_sub = data.frame()
  table_sub = subset(meta.data, celltype==Group[1])
  percentages=as.data.frame((table(table_sub$GLIPHAnnot)))$Freq
  table1 = percentages
  
  #m1 = meta.data
  
  for(j in 2:length(Group)){
    id = Group[j]
    n_total = table(meta.data$celltype)[[j]]
    totalNum=c(totalNum,n_total)
    
    table_sub = data.frame()
    table_sub = subset(meta.data, celltype==id)
    
    print(dim(table_sub))
    
    percentages=as.data.frame((table(table_sub$GLIPHAnnot)))$Freq
    table1=as.data.frame(rbind(table1,percentages))
    
    print(percentages)
  }
  
  table2 = as.data.frame(cbind(table1,Total=totalNum))
  rownames(table2)=Group
  colnames(table2)=c(levels(meta.data$GLIPHAnnot),"Total")
  
  table_cancer[,i] = table2[,1]/table2$Total
  table_noncancer[,i] = table2[,2]/table2$Total

}

outfile = paste(prefix,"TCRAnnot.GLIPH2Annot.Percent","Cancer","txt",sep=".")
write.table(table_cancer,file=outfile,sep="\t")

outfile = paste(prefix,"TCRAnnot.GLIPH2Annot.Percent","Viral","txt",sep=".")
write.table(table_noncancer,file=outfile,sep="\t")

## Enumerate the number of cancer and non-cancer related TCRs across all samples
meta.data = data.combined@meta.data
meta.data$GLIPHAnnot = factor(meta.data$GLIPHAnnot, levels = c("3-Cancer","1-Viral","2-AutoImmune"))
totalNum=c(table(meta.data$celltype)[[1]])
table_sub = data.frame()
table_sub = subset(meta.data, celltype==Group[1])
percentages=as.data.frame((table(table_sub$GLIPHAnnot)))$Freq
table1 = percentages
  
#m1 = meta.data
  
for(i in 2:length(Group)){
  id = Group[i]
  n_total = table(meta.data$celltype)[[i]]
  totalNum=c(totalNum,n_total)
    
  table_sub = data.frame()
  table_sub = subset(meta.data, celltype==id)
    
  #print(dim(table_sub))
    
  percentages=as.data.frame((table(table_sub$GLIPHAnnot)))$Freq
  table1=as.data.frame(rbind(table1,percentages))
    
  print(percentages)
}
  
table2 = as.data.frame(cbind(table1,Total=totalNum))
rownames(table2)=Group
colnames(table2)=c(levels(meta.data$GLIPHAnnot),"Total")

outfile = paste(prefix,"TCRAnnot.GLIPH2Annot.Percent","AllSamples","txt",sep=".")
write.table(table_noncancer,file=outfile,sep="\t")


```
