==> scRNAseq_Analysis/08_scRNAseq.Cellchat.Step2.CreatingSingleCellChatObject.rmd <==

#Load ALL CELLS from RDS file (and rename the clusters)
```{r}
library(Seurat)
library(Matrix)
library(dplyr)
library(ggplot2)
library(ggpubr)

RDSfile = "BrainMets.April2022/Seurat.MetsGBM.CombinedAnalysis.Apr2022.MinD0.1.sprd5.NN20.seed122.ep1000.rds"
data.combined = readRDS(RDSfile)

```

#create cellchat object for each condition
```{r}
library(CellChat)
library(patchwork)
options(stringsAsFactors = FALSE)

#assign subset annotation to each meta data seperately ###
##########################################################
filelist <- c("GBM",
              "GBM.pembro",
              "Mets",
              "Mets.pembro")

for(group in filelist){
  
  prefix = paste(group,"CellChat",sep = ".")
  
  t<-read.table(paste(group,"meta.txt", sep="."),sep="\t",header=T,row.names=1)
  
  ids <- rownames(t)
  subsetObject <- subset(data.combined, cells = ids)
  
  data.input = subsetObject@assays$RNA@data #normalized data matrix
  meta = subsetObject@meta.data
  
  newvals <- c("CD8-IL7R-CD69-Tm" = "CD8-Tm",
               "CD8-ITGA1-ITGAL-Trm" = "CD8-Trm.ex",
               "CD8-TCF7-CD226-Tprog.exh" = "CD8-Tprog.ex",
               "CD8-ISG High" = "CD8-Tem",
               "CD8-GZMK-CCL4-CCL5-Teff" = "CD8-Tearly.act.2",
               "CD8-GZMH-GZMA-CD52-Teff" = "CD8-Tearly.act.1",
               "CD8-VCAM1-IFNG-Tex" = "CD8-Tinter.ex",
               "CD8-CXCL13-LAG3-Tex" = "CD8-Tterm.ex",
               "CD4-IL7R-CCR7-CD40LG-Tcm" = "CD4-Tcm",
               "CD4-CXCL13-CTLA4-Tex" = "CD4-Tex",
               "Treg" = "Treg",
               "IL7R-KLRB1-Tm" = "CD4-CD8-Tm",
               "CTL/NK" = "CTL.NK",
               "CTL/NK-XCL1-XCL2" = "CTL.NK-XCL",
               "Lymphoid.Cycling" = "T.Cycling",
               
               #Myeloid 
               "Microglia" = "MG",
               "Microglia-ISG-IL10" = "MG-ISG",
               "Microglia-Phagocytosis" = "MG-Phagocytosis",
               "Monocyte" = "Mono",
               "Monocyte-ISG" = "Mono-ISG",
               "Mphage-ISG-MRC1-LILRB2" = "MDSC-ISG",
               "Mphage-Angiogenesis" = "Mphage-Angiogenesis",
               "Mphage-MRC1-LYVE1" = "Mphage-MRC1-LYVE1",
               "Mphage-Lysosome" = "Mphage-Lysosome",
               "Mphage-IL1B-CCL3-CCL4" = "Mphage-IL1B-CCL",
               "Complement/RiboHigh" = "Complement.Ribo",
               "Myeloid.Cycling" = "M.Cycling",                                     
               
               #DC             
               "cDC2" = "cDC2",
               "cDC1" = "cDC1",
               "pDC" = "pDC")
       
  meta$celltype2 = as.character(newvals[as.character(t$celltype)])
  meta$celltype2 = factor(meta$celltype2,levels = c("CD8-Tm",
                                                    "CD8-Tearly.act.1",
                                                    "CD8-Tearly.act.2",
                                                    "CD8-Tem",
                                                    "CD8-Trm.ex",
                                                    "CD8-Tprog.ex",
                                                    "CD8-Tinter.ex",
                                                    "CD8-Tterm.ex",
                                                    "CD4-Tcm",
                                                    "CD4-Tex",
                                                    "Treg",
                                                    "CD4-CD8-Tm",
                                                    "CTL.NK",
                                                    "CTL.NK-XCL",
                                                    "T.Cycling",
                                                    "MG",
                                                    "MG-ISG",
                                                    "MG-Phagocytosis",
                                                    "Complement.Ribo",
                                                    "Mono",
                                                    "Mono-ISG",
                                                    "MDSC-ISG",
                                                    "Mphage-Angiogenesis",
                                                    "Mphage-MRC1-LYVE1",
                                                    "Mphage-Lysosome",
                                                    "Mphage-IL1B-CCL",
                                                    "M.Cycling",
                                                    "cDC2",
                                                    "cDC1",
                                                    "pDC"))
  
  cellchat <- createCellChat(object = data.input, meta = meta, group.by = "celltype2")
  levels(cellchat@idents)
  
  groupSize <- as.numeric(table(cellchat@idents))
  print(groupSize)
  
  CellChatDB <- CellChatDB.human # use CellChatDB.mouse if running on mouse data
  #showDatabaseCategory(CellChatDB)
  
  CellChatDB.use <- CellChatDB
  cellchat@DB <- CellChatDB.use
  
  #Only take the genes in the database
  cellchat <- subsetData(cellchat)
  #future::plan("multiprocess", workers = 8, seed=TRUE)
  cellchat <- identifyOverExpressedGenes(cellchat)
  cellchat <- identifyOverExpressedInteractions(cellchat)
  cellchat <- projectData(cellchat, PPI.human)
  
  #Compute the communication probability and infer cellular communication network
  
  ##more strict computation of interaction probility - results in less nominated interaction
  cellchat <- computeCommunProb(cellchat,population.size = TRUE)
  
  #more lenient computation of interaction 
  #cellchat <- computeCommunProb(cellchat,population.size = TRUE,type="truncatedMean",trim=0.1)
  
  cellchat <- filterCommunication(cellchat, min.cells = 20)
  
  #Infer the cell-cell communication at a signaling pathway level
  cellchat <- computeCommunProbPathway(cellchat)
  cellchat <- aggregateNet(cellchat)
  signifPathway <- cellchat@netP$pathways
  
  saveRDS(cellchat,file=paste(prefix,"rds",sep="."))
  #saveRDS(cellchat,file=paste(prefix,"truncMean.rds",sep="."))
  
}

```
